requirements.txt
aniso8601==10.0.0
bcrypt==4.2.1
blinker==1.9.0
Brotli==1.1.0
cffi==1.17.1
chardet==5.2.0
click==8.1.8
colorama==0.4.6
cssselect2==0.7.0
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Navigation==0.2.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.55.4
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
pillow==11.1.0
pycparser==2.22
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.2
pytz==2025.1
reportlab==4.2.5
six==1.17.0
SQLAlchemy==2.0.37
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
webencodings==0.5.1
Werkzeug==3.1.3
wfastcgi==3.0.0
WTForms==3.2.1
zopfli==0.2.3.post1
